+++ lib/dockerfile.ml
@@ -257,4 +257,37 @@ module Linux = struct
       run "echo http://www.cl.cam.ac.uk/~avsm2/alpine-ocaml/%s >> /etc/apk/repositories" version @@
       install "ocaml camlp4"
   end
+
+  (* Zypper (opensuse) rules *)
+  module Zypper = struct
+    let update = run "zypper update -y"
+    let install fmt = ksprintf (fun s -> update @@ run "zypper install -y %s" s) fmt
+
+    let dev_packages ?extra () =
+      install "-t pattern devel_C_C++" @@
+      install "sudo git unzip curl" @@
+      (maybe (install "%s") extra)
+
+    let add_user ?uid ?gid ?(sudo=false) username =
+      let home = "/home/"^username in
+      run "useradd %s%s -d %s -m %s"
+        (match uid with None -> "" | Some d -> sprintf "-u %d " d)
+        (match gid with None -> "" | Some g -> sprintf "-g %d " g)
+        home username @@
+      (match sudo with
+        | false -> empty
+        | true ->
+         let sudofile = "/etc/sudoers.d/"^username in
+         run "echo '%s %s' > %s" username sudo_nopasswd sudofile @@
+         run "chmod 440 %s" sudofile @@
+         run "chown root:root %s" sudofile) @@
+       user "%s" username @@
+       workdir "%s" home @@
+       run "mkdir .ssh" @@
+       run "chmod 700 .ssh"
+
+     let install_system_ocaml =
+       install "ocaml camlp4"
+  end
+
 end